backend/requirements.txt
fastapi==0.115.6
google-genai>=0.10.0
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pymongo==4.10.1
dnspython==2.6.1
certifi==2025.1.31
passlib==1.7.4
python-jose[cryptography]==3.3.0
mediapipe>=0.10.32,<0.11
opencv-python>=4.12.0.88,<5
numpy>=2,<2.4
open-rppg
